config: Add CUE configuration schema
Initial outline of a configuration schema for the main daemon. CUE is
being used as the configuration format as it's pretty easy for humans to
read and write, and provides the ability for applying a schema like this
to validate the configuration. This should provide the ability to
validate configuration syntax outside of the daemon and by making the
validation part of the configuration format rather than implementing it
in code.