cmd/brd: Add the BorderRouter daemon
This is the "main" program of the OS. It reads the configuration file
and sets up networking.

If it's running as PID 1 (init) it will replaces most of gokrazy's init
functionality by calling the gokrazy functions to mount filesystems and
start services. In this mode any service configuration in gokrazy's
config isn't used: the service supervisor is currently only used to set
up the web interface and IP monitor.