cmd/brd: Mount overlay for /etc
/etc is part of the RO squashfs. We want to be able to write
/etc/resolv.conf and the system hostname based on the configuration
file.

We now use the /perm partition to create an overlay for /etc to be able
to write to it.

There's a catch with setting the system hostname: gokrazy uses a global
variable in gokrazy.Boot() to store the hostname, and this is used by
its status web interface, rather than the actual system hostname, so
once the hostname has been updated, the gokrazy status will show
whatever the system booted with (whatever was in /etc/hostname on the
root filesystem).